Refactor bridge command
